requirement.txt
Flask 
openai 
google-cloud-speech 
requests 
flask-cors 
python-dotenv 
gunicorn 
SpeechRecognition 
pyttsx3

